+++ src/backend/core/urls.py
@@ -56,5 +56,4 @@ urlpatterns = [
         ),
     ),
     path(f"api/{settings.API_VERSION}/config/", viewsets.ConfigView.as_view()),
-    path(f"api/{settings.API_VERSION}/footer/", viewsets.FooterView.as_view()),
 ]