+++ src/backend/impress/settings.py
@@ -423,19 +423,6 @@ class Base(Configuration):
         environ_name="FRONTEND_HOMEPAGE_FEATURE_ENABLED",
         environ_prefix=None,
     )
-    FRONTEND_URL_JSON_FOOTER = values.Value(
-        None, environ_name="FRONTEND_URL_JSON_FOOTER", environ_prefix=None
-    )
-    FRONTEND_FOOTER_FEATURE_ENABLED = values.BooleanValue(
-        default=False,
-        environ_name="FRONTEND_FOOTER_FEATURE_ENABLED",
-        environ_prefix=None,
-    )
-    FRONTEND_FOOTER_VIEW_CACHE_TIMEOUT = values.Value(
-        60 * 60 * 24,
-        environ_name="FRONTEND_FOOTER_VIEW_CACHE_TIMEOUT",
-        environ_prefix=None,
-    )
     FRONTEND_CSS_URL = values.Value(
         None, environ_name="FRONTEND_CSS_URL", environ_prefix=None
     )